+++ src/ethhelper/__init__.py
@@ -1,4 +1,8 @@
-from .connectors.http import GethHttpConnector
-from .connectors.ws import GethNewBlockSubscriber
+from .connectors.http import (
+    GethHttpConnector,
+)
+from .connectors.ws import (
+    GethNewBlockSubscriber,
+)
 
 __all__ = ["GethHttpConnector", "GethNewBlockSubscriber"]